MFC 359047,359054: Deprecate procfs-based process debugging.
359047:
Mark procfs-based process debugging as deprecated for FreeBSD 13.

Attempting to use ioctls on /proc/<pid>/mem to control a process will
trigger warnings on the console.  The <sys/pioctl.h> include file will
also now emit a compile-time warning when used from userland.

359054:
Fix the workaround to ignore the #warning for GCC.

clang and gcc use different warning flags for #warning preprocessor
directives.

For both 12 and 11, adjust the GCC warning flags to only be added in
4.7 and later since 4.2.1 does not support -Wno-cpp.  For 11, add the
needed warning suppression to procctl's build.  procctl was removed in
12.0.


git-svn-id: https://svn.freebsd.org/base/stable/11@361020 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f