MF11 361044,361048: Deprecate ubsec(4) for FreeBSD 13.0.
With the removal of in-tree consumers of DES, Triple DES, and
MD5-HMAC, the only algorithm this driver still supports is SHA1-HMAC.
This is not very useful as a standalone algorithm (IPsec AH-only with
SHA1 would be the only user).

This driver has also not been kept up to date with the original driver
in OpenBSD which supports a few more cards and AES-CBC on newer cards.
The newest card currently supported by this driver was released in
2005.

Approved by:	re (gjb)
Relnotes:	yes


git-svn-id: https://svn.freebsd.org/base/releng/11.4@361049 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f