MF11 r361047:
MFC r360915:

Use -fno-asynchronous-unwind-tables to compile lib/csu

Summary:
In r209294 kib added -fno-asynchronous-unwind-tables to the compile
flags for the GNU C startup components. This was done to work around a
BFD ld assertion, "no .eh_frame_hdr table will be created", which is
produced because of the layout of the startup objects.

Add the same flag to lib/csu too, for the same reason. And similarly to
r209294, also add -fno-omit-frame-pointer.

This is primarily meant to quickly MFC to stable/11, so it can end up in
the 11.4 release, as a fix for https://bugs.freebsd.org/246322.

Approved by:	re (gjb)
PR:		246322
Differential Revision: https://reviews.freebsd.org/D24797


git-svn-id: https://svn.freebsd.org/base/releng/11.4@361050 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f