MFC r360398: iflib: Stop interface before (un)registering VLAN
This is now misleadingly named for stable/11; this was meant to fix an
issue in an out-of-tree iavf(4) driver that uses iflib.

This commit actually introduces a new driver-dependent function that
iflib-using drivers can implement in order to tell iflib whether an
interface restart is needed for certain events. For this commit, this
function is only used for VLAN register/unregister events.

Sponsored by:	Intel Corporation


git-svn-id: https://svn.freebsd.org/base/stable/11@361055 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f