MFC r361250:
The -F flag of swapon(8) requires -a to work.

Before this change, swapon(8) implied that -F works as a standalone option,
which is not the case and would produce a usage message. This change extends
the description of the -F option to mention that -a is required with it.

PR:		238551
Submitted by:	Christian Baltini


git-svn-id: https://svn.freebsd.org/base/stable/12@361433 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f