MFC r361347: With RFC3168 ECN, CWR SHOULD only be sent with new data
Overly conservative data receivers may ignore the CWR flag on other
packets, and keep ECE latched. This can result in continous reduction
of the congestion window, and very poor performance when ECN is
enabled.

This does NOT contain the merge of the change to RACK since at this
time that code does not exist in stable/11, and there is no plan to
merge RACK to stable/11.

PR:		243590
Reviewed by:	rgrimes (mentor), rrs
Approved by:	rgrimes (mentor), tuexen (mentor)
MFC after:	3 days
Sponsored by:	NetApp, Inc.
Differential Revision:	https://reviews.freebsd.org/D23364


git-svn-id: https://svn.freebsd.org/base/stable/11@361436 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f