vfs: restore mtx-protected foffset locking for 32 bit platforms
They depend on it to accurately read the offset.

The new code is not used as it would add an interrupt enable/disable
trip on top of the atomic.

This also fixes a bug where 32-bit nolock request would still lock the offset.

No changes for 64-bit.

Reported by:	emaste


git-svn-id: https://svn.freebsd.org/base/head@361440 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f