MFS r361469: Fix bug in PR-SCTP Only drop DATA chunk with lower priorities as specified in RFC 7496. This issue was found by looking at a reproducer generated by syzkaller.
MFS r361472: Improve SCTP iterator
Ensure that the SCTP iterator runs with an stcb and inp, which belong to
each other.

MFS r361473: Improve stcb handling during teardown
Ensure that an stcb is not dereferenced when it is about to be freed.
This issue was found by SYZKALLER.

MFS r361476: Improve ASCONF handling
Avoid an integer underflow.

Approved by:	re(gjb)


git-svn-id: https://svn.freebsd.org/base/releng/11.4@361522 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f